filtring is ready
