+++ src/view/event-with-content-view.js
@@ -1,7 +1,7 @@
 import { RenderPosition, createElement } from '../framework/render';
 import { MOVING_ELEMENTS, mapCitys, mapOffers } from '../mocks/mock';
 import AbstractStatefulView from '../framework/view/abstract-stateful-view';
-import { INPUT } from '../framework/conts';
+import { INPUT } from '../framework/consts';
 import flatpickr from 'flatpickr';
 import 'flatpickr/dist/flatpickr.min.css';
